Updating Serving to handle null/empty addresses
